Moved memory db initialize code to test module's preinitialize.
